Extract UIImage helper to a separate file
